+++ src/analysis/multi_task_writes.mli
@@ -1,0 +1,6 @@
+(** Detect outputs, memory and global variables written by program instances
+    that run in different tasks. *)
+open IECCheckerCore
+module S = Syntax
+
+val run : S.iec_library_element list -> Warn.t list